+++ src/search_params.hpp
@@ -45,9 +45,9 @@ public:
 
 }; // struct TunableParam
 
-constexpr i32 MAX_DEPTH = 100;
+constexpr size_t MAX_DEPTH = 100;
 constexpr i32 INF = 30'000;
-constexpr i32 MIN_MATE_SCORE = INF - MAX_DEPTH;
+constexpr i32 MIN_MATE_SCORE = INF - static_cast<i32>(MAX_DEPTH);
 
 // Time management
 MAYBE_CONSTEXPR TunableParam<double> hardTimePercentage = TunableParam<double>(0.5, 0.25, 0.75, 0.05);
